Add beam-sql99 and beam-sql2003
