+++ beam-core/Database/Beam/Query/Operator.hs
@@ -1,14 +1,12 @@
-module Database.Beam.Query.Operator where
-  -- ( SqlBool
+module Database.Beam.Query.Operator
+  ( SqlBool
 
-  --   -- ** General-purpose operators
-  -- , (&&.), (||.), not_, div_, mod_
-  -- , (&&?.), (||?.), sqlNot_
+    -- ** General-purpose operators
+  , (&&.), (||.), not_, div_, mod_
+  , (&&?.), (||?.), sqlNot_
 
-  -- , like_, similarTo_
-
-  -- , concat_
-  -- ) where
+  , like_, similarTo_
+  ) where
 
 import           Database.Beam.Syntax
 import           Database.Beam.Backend.SQL
@@ -55,24 +53,18 @@ data SqlBool
        -> QGenExpr context s SqlBool
 (||?.) = qBinOpE orE
 
--- infixr 3 &&., &&?.
--- infixr 2 ||., ||?.
+infixr 3 &&., &&?.
+infixr 2 ||., ||?.
 
--- -- | SQL @LIKE@ operator
--- like_ ::
---   ( IsSqlExpressionSyntaxStringType syntax text
---   , IsSql92ExpressionSyntax syntax ) =>
---   QExpr syntax s text -> QExpr syntax s text -> QExpr syntax s Bool
--- like_ (QExpr scrutinee) (QExpr search) =
---   QExpr (liftA2 likeE scrutinee search)
+-- | SQL @LIKE@ operator
+like_ :: QExpr s text -> QExpr s text -> QExpr s Bool
+like_ (QExpr scrutinee) (QExpr search) =
+  QExpr (liftA2 likeE scrutinee search)
 
--- -- | SQL99 @SIMILAR TO@ operator
--- similarTo_ ::
---   ( IsSqlExpressionSyntaxStringType syntax text
---   , IsSql99ExpressionSyntax syntax ) =>
---   QExpr syntax s text -> QExpr syntax s text -> QExpr syntax s text
--- similarTo_ (QExpr scrutinee) (QExpr search) =
---   QExpr (liftA2 similarToE scrutinee search)
+-- | SQL99 @SIMILAR TO@ operator
+similarTo_ :: QExpr s text -> QExpr s text -> QExpr s text
+similarTo_ (QExpr scrutinee) (QExpr search) =
+  QExpr (liftA2 similarToE scrutinee search)
 
 -- | SQL @NOT@ operator
 not_ :: ExpressionContext context
@@ -86,19 +78,14 @@ sqlNot_ :: ExpressionContext context
         -> QGenExpr context s SqlBool
 sqlNot_ (QExpr a) = QExpr (fmap notE a)
 
--- -- | SQL @/@ operator
--- div_ :: (Integral a, IsSql92ExpressionSyntax syntax)
---      => QGenExpr context syntax s a -> QGenExpr context syntax s a
---      -> QGenExpr context syntax s a
--- div_ = qBinOpE divE
+-- | SQL @/@ operator
+div_ :: (Integral a, ExpressionContext context)
+     => QGenExpr context s a -> QGenExpr context s a
+     -> QGenExpr context s a
+div_ = qBinOpE divE
 
--- -- | SQL @%@ operator
--- mod_ :: (Integral a, IsSql92ExpressionSyntax syntax)
---      => QGenExpr context syntax s a -> QGenExpr context syntax s a
---      -> QGenExpr context syntax s a
--- mod_ = qBinOpE modE
-
--- -- | SQL @CONCAT@ function
--- concat_ :: IsSql99ConcatExpressionSyntax syntax
---         => [ QGenExpr context syntax s T.Text ] -> QGenExpr context syntax s T.Text
--- concat_ es = QExpr (concatE <$> mapM (\(QExpr e) -> e) es)
+-- | SQL @%@ operator
+mod_ :: (Integral a, ExpressionContext context)
+     => QGenExpr context s a -> QGenExpr context s a
+     -> QGenExpr context s a
+mod_ = qBinOpE modE
