+++ beam-core/Database/Beam/Query/Types.hs
@@ -1,5 +1,5 @@
 module Database.Beam.Query.Types
-  ( Q, QExpr, QGenExpr(..), QExprToIdentity, QExprToField, QWindow, QWindowFrame
+  ( Q, QExpr, QGenExpr(..), QExprToIdentity, QExprToField
 
   , Projectible, Aggregation
