+++ beam-core/beam-core.cabal
@@ -42,7 +42,6 @@ library
                        Database.Beam.Backend.SQL.BeamExtensions
   other-modules:       Database.Beam.Query.Aggregate
                        Database.Beam.Query.Combinators
-                       Database.Beam.Query.Extensions
                        Database.Beam.Query.Operator
                        Database.Beam.Query.Ord
                        Database.Beam.Query.Relationships
