+++ beam-sql99/beam-sql99.cabal
@@ -1,0 +1,10 @@
+name: beam-sql99
+version: 0.1
+cabal-version: 2.0
+build-type: Simple
+            
+library
+    build-depends: base >= 4.8, beam-core
+    signatures: Database.Beam.Syntax
+    exposed-modules: Database.Beam.Query.SQL99
+    default-language: Haskell2010
